Rebuilt dev environment (#3)
Previously: Developping live required to launch the demo-app present in the /demo directory. It loaded react-native-masonry as a local dependency to make sure the latest version was used.

Problem: npm was copying files instead of linking them, meaning any change made in /lib weren't reflected in /demo. The only was was to uninstall / reinstall the package each time (and restart the packager to see the changes).

Fix: The/demo app (renamed to MasonryExample) is now part of the root directory, and therefore has direct access to the library source files without needing to require it with package.json. This makes developping the lib and previewing the changes in simulator / device work naturally. The files are excluded for npm publication so don't add to the weight of the final package.